% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLdd.R
\name{LLdd}
\alias{LLdd}
\title{loglikelihood for model}
\usage{
LLdd(hr, thetaX1, thetaY1, betaXdy, betaXdy2, betaXhr, betaXhr2, betaYdy,
  betaYdy2, betaYhr, betaYhr2, delta, phi, kappa1, sigX, sigY, d2, H, adelta,
  bdelta, aphi, bphi, akappa, bkappa, psi, mupsi, xi, rho)
}
\arguments{
\item{hr}{hour of day}

\item{thetaX1}{the starting value for the X-variable, which is the underlying pup population}

\item{thetaY1}{the starting value for the Y-variable, which is the underlying nonpup population}

\item{betaXdy}{day-of-year linear regression coefficient for pups}

\item{betaXdy2}{day-of-year quadratic regression coefficient for pups}

\item{betaXhr}{hour-of-day linear regression coefficient for pups}

\item{betaXhr2}{hour-of-day quadratic regression coefficient for pups}

\item{betaYdy}{day-of-year linear regression coefficient for nonpups}

\item{betaYdy2}{day-of-year quadratic regression coefficient for nonpups}

\item{betaYhr}{hour-of-day linear regression coefficient for nonpups}

\item{betaYhr2}{hour-of-day quadratic regression coefficient for nonpups}

\item{delta}{annual survivorship of nonpups}

\item{phi}{annual fecundity of nonpups to time of survey}

\item{kappa1}{annual survivorship of pups from time to survey to yearling}

\item{sigX}{measurement error for pups}

\item{sigY}{measurement error for nonpups}

\item{d2}{dataset}

\item{H}{harvest dataset}

\item{adelta}{beta parameter a for prior on delta}

\item{bdelta}{beta parameter b for prior on delta}

\item{aphi}{beta parameter a for prior on phi}

\item{bphi}{beta parameter b for prior on phi}

\item{akappa}{beta parameter a for prior on kappa}

\item{bkappa}{beta parameter b for prior on kappa}

\item{psi}{proportion of animals that are out of the water}

\item{xi}{exponential scale parameter for prior on psi}

\item{rho}{density dependent fecundity parameter}
}
\value{
the loglikelihood
}
\description{
evaluates to the loglikelihood for MCMC sampling
}
\author{
Jay Ver Hoef
}
