% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCddPsiUni.R
\name{MCMCddPsiUni}
\alias{MCMCddPsiUni}
\title{MCMC sampling}
\usage{
MCMCddPsiUni(burnin = 10000, niter = 1e+05, thin = 100, adelta, bdelta,
  aphi, bphi, akappa, bkappa, harvestData, data, psi.lower, psi.upper, xi,
  maxSight)
}
\arguments{
\item{burnin}{the number of MCMC for burnin}

\item{niter}{the number of MCMC iterations}

\item{thin}{thinning of the MCMC chain}

\item{adelta}{beta parameter a for prior on delta}

\item{bdelta}{beta parameter b for prior on delta}

\item{aphi}{beta parameter a for prior on phi}

\item{bphi}{beta parameter b for prior on phi}

\item{akappa}{beta parameter a for prior on kappa}

\item{bkappa}{beta parameter b for prior on kappa}

\item{harvestData}{harvest dataset}

\item{data}{survey dataset}

\item{psi.lower}{lower bound on uniform prior for psi}

\item{psi.upper}{upper bound on uniform prior for psi}

\item{maxSight}{the maximum sightability across all surveys}
}
\value{
a list of the MCMC chains values
}
\description{
MCMC sampling
}
\author{
Jay Ver Hoef
}
