% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCddU.R
\name{MCMCddU}
\alias{MCMCddU}
\title{MCMC sampling}
\usage{
MCMCddU(burnin = 10000, niter = 1e+05, thin = 100, delta.lower = 0.5,
  delta.upper = 1, phi.lower = 0, phi.upper = 0.5, kappa.lower = 0.5,
  kappa.upper = 1, harvestData, data, mupsi, xi, maxSight)
}
\arguments{
\item{burnin}{the number of MCMC for burnin}

\item{niter}{the number of MCMC iterations}

\item{thin}{thinning of the MCMC chain}

\item{delta.lower}{lower bound for uniform prior on delta}

\item{delta.upper}{upper bound for uniform prior on delta}

\item{phi.lower}{lower bound for uniform prior on phi}

\item{phi.upper}{upper bound for uniform prior on phi}

\item{kappa.lower}{lower bound for uniform prior on kappa}

\item{kappa.upper}{upper bound for uniform prior on kappa}

\item{harvestData}{harvest dataset}

\item{data}{survey dataset}

\item{xi}{exponential scale parameter for prior on psi}

\item{maxSight}{the maximum sightability across all surveys}
}
\value{
a list of the MCMC chains values
}
\description{
MCMC sampling
}
\author{
Jay Ver Hoef
}
